Only use one Cipher object
There's no need for a second Cipher. Since the payload for the email verification link is unique anyway (it has items like "withdrawal_reason" that aren't in other tokens), this means the generated token will also be unique. This is all that's required for security, since everything is encrypted. No need for 2 ciphers.