Don't show 404s for errors
Showing a 404 when something has gone wrong in the code is hopelessly confusing for both the user and the developer.

I've made it so most errors now bubble up to create standard 500 pages, which will be much more straightforward for everyone to reason about. The only special case I've added is that if the candidate_id doesn't match I'm throwing a 403 instead, as that's probably closer to the truth.