Move templates; Withdrawal requested page; Improve debug messaging
Moved templates to fit with the URL structure (which is our standard pattern).
---

As in, template for canonical.com/careers/application is at
templates/careers/application/index.html

Make a dedicated page for "withdrawal requested"
---

Since requesting a withdrawal requires submitting a form to the backend,
which then reloads the page completely, it's unnecessary and clunky
to redirect back to the index page with a `#` target.

It's much more straightforward to simply show a dedicated "withdrawal requested"
confirmation page. And also this enables my debug messaging described below:

Improve debug behaviour for sending email
---

Only skip sending the email if we know we're in debug mode.

If you're not in debug mode, the `send_mail` function will expect
the appropriate environment variables (SMTP_SERVER, etc.) to be defined.
If they're not defined, a standard KeyError will be raised, leading to a server
error, to make it very clear that configuration settings are missing.

If you're in debug mode, a message will be printed to the