Replace withdrawal notification
The reason that we wanted the "withdrawal requested" message to be displayed
as a notification rather than as a dedicated page was because the dedicated
page could mislead the user (who may only skim it) into thinking the withdrawal
has already gone through. We're seeking to avoid that by making it clear
that their application is still active by showing them the application page.

However I still think the redirect flow was clunky. So instead I'm processing
the withdrawal at the same URL as the application index, and returning
the application index page directly back, without a redirect. This allows
me to put the debug message in place as I was before.

I've also tidied up the code a fair bit, moving shared code into helper
functions etc.