WD-1163: Turn vacancy.departments into a proper array
Previously "departments" was usually stored as a string containing only the first department.

This led to the bug in WD-1163, as in one context it was looking for jobs where
the department was contained within the departments array, but in other contexts
only the first department was represented.

This fix makes it so all departments for a vacancy are considered on all pages.